test(profile): refactor import paths and remove obsolete profile tests
